requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
aiofiles
pillow
openai
python-dotenv
requests
numpy